🥇 "in honor of" (#46)
* removed a :

removing the colon (again!)

(testing was a good idea, since this was the one file I didn't test before. because I can see locally that it *should* be right again.. but I then have to avoid checking in package-lock and yarn - again! - which should be happening...now...)

* sigs overview and..

"in honor of"

* SIGs and .

Co-authored-by: Chase 李 <clee231@uic.edu>

Co-authored-by: Chase 李 <clee231@uic.edu>